TextGraphs

Documentation for TextGraphs.

TextGraphs.add_prop_label_tokens

TextGraphs.build_labelled_graph

TextGraphs.lemma_graph

TextGraphs.link_consecutive

TextGraphs.naive_graph

TextGraphs.phrases_graph

TextGraphs.pos_graph

TextGraphs.stem_graph

TextGraphs.udp_import_annotations

TextGraphs.add_prop_label_tokens — Method

add_prop_label_tokens(metagraph,metagraph_unique_tokens)

Add tokens as properties of nodes in a MetaGraph.

This function is used internally to attach word labels to each node. Unique tokens must have length equal to the number of vertices

TextGraphs.build_labelled_graph — Method

build_labelled_graph(x::AbstractArray)

This function is used internally to build graph lebelled with unique tokens.

TextGraphs.lemma_graph — Method

lemma_graph(my_text)

Build lemmatized graph from text (AbstractString) using R package udpipe.

Currently, supports portuguese and english corpora.

TextGraphs.link_consecutive — Method

link_consecutive(array_with_tokens)

Transform serialized tokens into a directed graph.

This function is used internally to build graphs from text. Each token has an unique node in the graph.

TextGraphs.naive_graph — Method

naive_graph(raw_text::AbstractString)

Build graph from text (AbstractString) with unprocessed words.

TextGraphs.phrases_graph — Method

phrases_graph(my_text)

Build graph from text (AbstractString) using sentences as unique tokens.

TextGraphs.pos_graph — Method

pos_graph(my_text)

Build POS Tagging from text (AbstractString) using R package udpipe.

Currently, supports portuguese and english corpora.

TextGraphs.stem_graph — Method

stem_graph(my_text)

Build graph from text (AbstractString) using lemmatized words.

Stemming is performed with Snowball.jl stemmer. Default language is "portugese".

TextGraphs.udp_import_annotations — Method

udp_import_annotations(raw_text)

Gets anonnotated DataFrame by importing R::udpipe object created with udpipe::annonate.

This function is used internally.